Create Biohunter.py
A LOT of tweaking needed.  Eventually, I want to move these methods into a "searchagent" class, a "dataprocessor" class, and a "filingclerk" class so that we can have search+download and process+commit-to-mysql be separate and parallel processes.  Imagination time: Four(or however many) search-agents are given a list of author names by the filing-clerk (from mysql) to search for and retrieve from ANB and other websites.  They then deposit the information the find in the queue for the x-number of data-processors to sift through.  The information extracted is then handed over to the filing clerk to put back in mysql.  This allows the data-processors to start working their magic as soon as information starts coming in rather than having to wait for everything to be downloaded, and to prevent a failed download of one item from FUBAR the whole thing.