Create NERuntokenizer.py
In the end, this turned out to be just an unnecessary programming exercise, as the Stanford NER testing actually required the tab separated format that the gold standard was already in in the first place.  oops.  So much for reading the directions before diving in.  In any case, this takes the tsv files and turns them back into raw unformatted text.